rust: emit insert patches for block markers
Problem: when receiving changes over the network the logic for the
patches emitted was incorrectly emitting "splice" patches for block
markers.

Solution: emit "insert" patches for block markers instead. This patch
actually covers two cases in the code where this could happen but one of
them is not possible to test as it can't actually be triggered in
practice. This case is the case when an op received over the network
updates an index in a text sequence to be a block marker. At the moment
the API doesn't allow us to emit update operations for text sequences.