Fix splice deleting marks which are on the boundaries of the splice
Problem: using Automerge::splice_text to replace a range of text which
has marks on the boundaries of the splice would result in the marks
being removed, even if the marks were set to expand. The reason this
happens is because splice is implemented as a deletion followed by an
insertion, which means that when we insert the new text the marks are
empty and so the newly inserted text can't be inside the marks.

For example, say we have this text (where <bold> indicates a bold mark,
which is set to expand on both sides):

    <bold>hello</bold> world

If we call `Automerge::splice_text(0, 5, "hi")` to replace "hello" we
would end up with:

    hi world

Instead, we want to end up with:

    <bold>hi</bold> world


Solution: reorder the splice operation to first insert the new text,
then delete the old text. This means that the new text takes account of
mark expansion.

In order to achieve this I needed to add a call to `seek_ops_by_index`
at the start of the `splice_text` method in order to determine if the
insertion is into a multibyte character. This in turn caused a few
errors due to the `OpSet::seek_ops_by_index_slow` method returning
`OpsFound` with different ranges on it to the
`OpSet::seek_text_ops_by_index_fast` method. I've normalised the ranges
returned in these cases.

Ideally we would actually return the index as part of the `QueryNth`
returned by `OpSet::query_insert_at` function in order to avoid doing an
extra query on every splice. Unfortunately this seems to be quite
tricky. In the `query_insert_at_text` method we seek to the
insertion offset using `Indexes::text::iter_range::with_acc` followed by
`nth` which seeks to the given offset in terms of character widths. This
allows us to tell if the index is different to the offset indicating a
multibyte character somewhere in the text, but we can't figure out what
the width of the character we are currently pointing at is.