Fix splice deleting marks which are on the boundaries of the splice
Problem: using Automerge::splice_text to replace a range of text which
has marks on the boundaries of the splice would result in the marks
being removed, even if the marks were set to expand. The reason this
happens is because splice is implemented as a deletion followed by an
insertion, which means that when we insert the new text the marks are
empty and so the newly inserted text can't be inside the marks.

For example, say we have this text (where <bold> indicates a bold mark,
which is set to expand on both sides):

    <bold>hello</bold> world

If we call `Automerge::splice_text(0, 5, "hi")` to replace "hello" we
would end up with:

    hi world

Instead, we want to end up with:

    <bold>hi</bold> world


Solution: reorder the splice operation to first insert the new text,
then delete the old text. This means that the new text takes account of
mark expansion.

In order to achieve this I initially used a call to `seek_ops_by_index`
at the start of the `splice_text` method in order to determine if the
insertion is into a multibyte character. This in turn caused a few
errors due to the `OpSet::seek_ops_by_index_slow` method returning
`OpsFound` with different ranges on it to the
`OpSet::seek_text_ops_by_index_fast` method. I've normalised the ranges
returned in these cases.

Orion later figured out how to include the index in `QueryNth` so we
don't do the `seek_ops_by_index` call at all.

Co-authored-by: Orion Henry <orion.henry@gmail.com>