rust: Remove TextRepresentation
Until the recent 3.0 release of the automerge JavaScript library we
supported two representations of collaborative text. One as a list of
characters, and another as a native JavaScript string. In order to
support both representations the rust implementation has a
`TextRepresentation` enum which specifies how text should be represented
in patches. This type has spread throughout the codebase and is used for
multiple purposes:

* Determining what text encoding to use when measuring the width of
  characters in a document
* Determining how to serialize text in patches

Now that we no longer support the list representation of text we can
remove this type. While here I also did the following:

* Rename the ListEncoding type to SequenceType, which captures its use
  more precisely
* Remove the nested `TextEncoding` enum from the `SequenceType::Text`
  enum variant. This means that code which doesn't care about the text
  encoding doesn't have to construct a default value for it.
* Add `ObjType::as_sequence_type() -> Option<SequenceType>` to determine
  the sequence type of an object.