rust: Do not implement Default for TextEncoding
Implementing Default for TextEncoding is error prone as it makes it easy
to accidentally use the default encoding in derived `Default`
implementations. Instead of implementing default this commit switches to
providing a method called `TextEncoding::platform_default` for code
paths which explicitly want to use a default and the rest of the time
requires that the caller pass the encoding explicitly. This makes it
harder to accidentally use a different encoding to the encoding the
OpSet is using.