rust fix empty changes not being filtered out of get_changes
Problem: creating an empty change and then calling
`Automerge::get_changes(<hash of empty change>)` would return a vector
containing the empty change, implying that the empty change is not
contained within it's own heads. This leads to a number of infinite
loops when syncing with empty documents.

Why was this happening? In get_changes we determine which changes to
filter out (i.e. which changes are not ancestors of the given heads) by
walking the commit graph up to the heads we were given and creating a
vector clock of all the changes we saw on the way. This clock is then
used to filter the overall change graph. When we create this clock we
were deciding whether to incorporate the sequence number of a change
into the clock by comparing the max_op of the change with the largest
max_op we've seen so far. In the case of an empty change the max_op
doesn't increase, so we were never incorporating the larger sequence
number.

Solution: Always incorporate the sequence number, even if the max_op
hasn't changed.