Fix a clock caching bug
This fixes a particularly pernicious bug which caused basically any API
which accepst a set of heads (e.g. the sync protocol, the various *_at
methods, the Automerge.view API in JS and so on) to return nonsensical
results.

The reason this problem arose is to do with the way that automerge uses
change hashes to identify points in time internally. Every commit in an
Automerge document has an (actor ID, sequence number) pair (a lamport
timestamp) which identifies the commit. We can walk the commit graph
accumulating these pairs to form a vector clock. This vector clock can
then be used to determine if any given commit is an ancestor of the
given heads. To identify the state of a document at a given set of heads
then, we compute the vector clock from the commit graph for those heads.

E.g. imagine this graph where I denote the lamport timestamp as (actor
ID, sequence number):

                      * (a, 1)
                      |
                      * (a, 2)
                    /   \
                   /      \
                  * (b, 1) * (c, 1)
                   \      /
                    \   /
                      * (d, 1)

Then say we have the heads for (b, 1), we can walk from (b,1) to the
root and we will end up with a vector clock of [(a, 2), (b, 1)]. We can
then use this vector clock to determine that (b,1) is an ancestor of
(a,1) because the clock for (b,1) contains (a,2), which is greater than
(a,1)

Creating these clocks for large documents is expensive, so we cache
clocks for every 16 commits in the document. The problem this commit
fixes is that the logic for the clock caching is incorrect. When walking
the graph, the cache logic can accidentally omit some commits from the
cached clock.

The caching logic is expressed by starting with the end node and then
walking backwards in a depth first search until the cache limit is
reached, it looks something like this:

let limit = CACHE_STEP * 2
let clock = new_clock()
let to_visit = [start node]
let visited = []
while let Some(node) = to_visit.pop() {
    // Process node
    if let Some(cached) = get_cache(node) {
        merge(cached, clock)
    } else if visited.len() <= limit {
        to_visit.extend(parents_of(node))
    } else {
        break;
    }
}

The logic will then restart from the remaining nodes in the to_visit queue.

The bug is that in some scenarios we don't add the parents of the node
which causes us to reach the cache limit to the to_visit queue.

The easiest way to observe this bug is to create a document where there
is a commit which has a large number of parents (i.e. a merge commit
from many branches). E.g imagine a document like this:

                    A
                    |
        .--.--.--.--.--.--.--.--.
        B  C  D  E  F  G  H  I  J
        |  |  |  |  |  |  |  |  |
        K  L  M  N  O  P  Q  R  S
        '--'--'--'--'--'--'--'--'
                   |
                   T
                   |
                   U

Now, let's say our CACHE_STEP is 3. Then every third change which is
applied is cached. U is the 21st change and so it will be cached. The
cache logic will now step backwards through the graph until 6 (CACHE_STEP
* 2) nodes have been visited, and then it will stop. In this case that
would mean that we add T to the list of nodes to visit, then pop it and
process it, adding all of T's parents to the list of nodes to visit.
Then we continue processing until we reach O so the final order is
T, K, L, M, N, O. Here's where the bug is, the cache logic now returns
and says "now keep going with the rest of the to_visit list", but it
didn't add the parents of O to the list and so F never gets processed.

This is only a problem if F is a commit with a different actor ID then
O, otherwise the (actor_id, seq) if O covers F.

The fix is very simple, belying the complexity of chasing this down.

Co-authored-by: Orion Henry <orion.henry@gmail.com>