rust: update to rustc 1.89
This is primarily necessary because smol_str has documented their
requirement for rustc 1.89 as of smol_str 0.3.4. This is not a new
requirement though, it was always necessary and so we also need to
upgrade.

There's a bunch of new clippy lints as a result of the upgrade which
I've fixed here as well.