rust: only hold a reference to Columns in OpIter
In order to implement interruptible loading we need to be able to
suspend iterators over the opset. This will require some unsafe code as
we are holding references to the data of the opset columns across yield
points. Currently the OpIter holds a reference to the entire OpSet, but
it only really needs the columns (i.e. OpSet::columns). In order to make
it easier to reason about soundness of the unsafe code we are going to
need to write, change OpIter to only hold a reference to Columns.