rust: Add ReadDoc::view_at
Problem: sometimes (especially in generic code) you want to work
generically with `ReadDoc`, but the caller wants to be able to specify
the head to view at. To achieve that at the moment you have to accept an
`Option<&[ChangeHash]>` parameter and then call the appropriate
`ReadDoc::*_at` method if it's `Some` or the normal method if it's
`None`. This is awkward and relies on the implementor to add this logic.

Solution: add a `ReadDoc::view_at(&self, heads: Option<&[ChangeHash]>)`
method that returns an implementation of `ReadDoc` that returns the
values from the original `ReadDoc` but viewed at the specified heads.